Fix: Move switch logics to separated module
