+++ node/switch.py
@@ -1,4 +1,7 @@
 import asyncio
+import json
+import threading
+import time
 from typing import Any
 
 import aiohttp
@@ -8,6 +11,8 @@ from common.batch import stateful_batch_to_batch_record
 from common.db import zdb
 from common.logger import zlogger
 from config import zconfig
+
+switch_lock: threading.Lock = threading.Lock()
 
 
 async def find_all_nodes_last_finalized_batch_record_async(app_name: str) -> BatchRecord:
@@ -90,3 +95,89 @@ async def _find_all_nodes_last_finalized_batch_record_core(app_name: str) -> Bat
                 highest_record = record
 
     return highest_record
+
+
+async def send_switch_request(session, node, proofs):
+    """Send a single switch request to a node."""
+    data = json.dumps({
+        "proofs": proofs,
+        "timestamp": int(time.time()),
+    })
+    url = f'{node["socket"]}/node/switch'
+
+    try:
+        async with session.post(url, data=data, headers=zconfig.HEADERS) as response:
+            await response.text()  # Consume the response
+    except Exception as error:
+        zlogger.error(f"Error occurred while sending switch request to {node['id']}")
+
+
+async def send_switch_requests(proofs: list[dict[str, Any]]) -> None:
+    """Send switch requests to all nodes except self asynchronously."""
+    zlogger.warning("sending switch requests...")
+    async with aiohttp.ClientSession() as session:
+        tasks = [
+            send_switch_request(session, node, proofs)
+            for node in zconfig.NODES.values() if node["id"] != zconfig.NODE["id"]
+        ]
+        await asyncio.gather(*tasks)
+
+
+def verify_switch_sequencer(old_sequencer_id: str) -> bool:
+    return old_sequencer_id == zconfig.SEQUENCER["id"]
+
+
+def switch_sequencer(old_sequencer_id: str, new_sequencer_id: str):
+    """
+    Synchronous wrapper for sequencer switching.
+    """
+    with switch_lock:
+        asyncio.run(_switch_sequencer_core(old_sequencer_id, new_sequencer_id))
+
+
+async def switch_sequencer_async(old_sequencer_id: str, new_sequencer_id: str):
+    """
+    Asynchronous version of sequencer switching.
+    """
+    with switch_lock:
+        await _switch_sequencer_core(old_sequencer_id, new_sequencer_id)
+
+
+async def _switch_sequencer_core(old_sequencer_id: str, new_sequencer_id: str):
+    """
+    Core implementation of sequencer switching logic.
+    Used by both sync and async switch functions.
+    """
+    if not verify_switch_sequencer(old_sequencer_id):
+        return
+
+    zdb.pause_node.set()
+
+    try:
+        zconfig.update_sequencer(new_sequencer_id)
+
+        tasks = [
+            sync_app_finalized_state(app_name, new_sequencer_id)
+            for app_name in list(zconfig.APPS.keys())
+        ]
+        await asyncio.gather(*tasks)
+
+        if zconfig.NODE['id'] != zconfig.SEQUENCER['id']:
+            await asyncio.sleep(10)
+    finally:
+        zdb.pause_node.clear()
+
+
+async def sync_app_finalized_state(app_name: str, new_sequencer_id: str):
+    """
+    Process a single app during sequencer switch by updating its finalized batch records
+    and resetting batch timestamps.
+
+    Args:
+        app_name: Name of the app to process
+        new_sequencer_id: ID of the new sequencer
+    """
+    all_nodes_last_finalized_batch_record = await find_all_nodes_last_finalized_batch_record_async(app_name)
+    if all_nodes_last_finalized_batch_record:
+        zdb.reinitialize(app_name, new_sequencer_id, all_nodes_last_finalized_batch_record)
+    zdb.reset_not_finalized_batches_timestamps(app_name)
