+++ node/tasks.py
@@ -6,7 +6,6 @@ dispute handling, and switching sequencers.
 import asyncio
 import json
 import random
-import threading
 import time
 from typing import Any
 from typing import List
@@ -20,11 +19,9 @@ from common.db import zdb
 from common.errors import ErrorCodes, ErrorMessages
 from common.logger import zlogger
 from config import zconfig
-from node.node_queries_task import find_all_nodes_last_finalized_batch_record_async
 from node.rate_limit import get_remaining_capacity_kb_of_self_node
 from node.rate_limit import try_acquire_rate_limit_of_self_node
-
-switch_lock: threading.Lock = threading.Lock()
+from node.switch import send_switch_requests, switch_sequencer_async
 
 
 def check_finalization() -> None:
@@ -323,7 +320,8 @@ async def send_dispute_requests() -> None:
     if is_not_synced or (no_missed_batches and sequencer_up) or is_paused:
         return
 
-    zlogger.warning(f"Sending dispute is_not_synced: {is_not_synced}, no_missed_batches: {no_missed_batches}, sequencer_up: {sequencer_up}, is_paused: {is_paused}")
+    zlogger.warning(
+        f"Sending dispute is_not_synced: {is_not_synced}, no_missed_batches: {no_missed_batches}, sequencer_up: {sequencer_up}, is_paused: {is_paused}")
     timestamp: int = int(time.time())
     new_sequencer_id: str = utils.get_next_sequencer_id(
         old_sequencer_id=zconfig.SEQUENCER["id"]
@@ -382,89 +380,3 @@ async def send_dispute_request(
                     return response_json.get("data")
     except aiohttp.ClientError as error:
         zlogger.warning(f"Error sending dispute request to {node['id']}: {error}")
-
-
-async def send_switch_request(session, node, proofs):
-    """Send a single switch request to a node."""
-    data = json.dumps({
-        "proofs": proofs,
-        "timestamp": int(time.time()),
-    })
-    url = f'{node["socket"]}/node/switch'
-
-    try:
-        async with session.post(url, data=data, headers=zconfig.HEADERS) as response:
-            await response.text()  # Consume the response
-    except Exception as error:
-        zlogger.error(f"Error occurred while sending switch request to {node['id']}")
-
-
-async def send_switch_requests(proofs: list[dict[str, Any]]) -> None:
-    """Send switch requests to all nodes except self asynchronously."""
-    zlogger.warning("sending switch requests...")
-    async with aiohttp.ClientSession() as session:
-        tasks = [
-            send_switch_request(session, node, proofs)
-            for node in zconfig.NODES.values() if node["id"] != zconfig.NODE["id"]
-        ]
-        await asyncio.gather(*tasks)
-
-
-def verify_switch_sequencer(old_sequencer_id: str) -> bool:
-    return old_sequencer_id == zconfig.SEQUENCER["id"]
-
-
-async def _switch_sequencer_core(old_sequencer_id: str, new_sequencer_id: str):
-    """
-    Core implementation of sequencer switching logic.
-    Used by both sync and async switch functions.
-    """
-    if not verify_switch_sequencer(old_sequencer_id):
-        return
-
-    zdb.pause_node.set()
-
-    try:
-        zconfig.update_sequencer(new_sequencer_id)
-
-        tasks = [
-            sync_app_finalized_state(app_name, new_sequencer_id)
-            for app_name in list(zconfig.APPS.keys())
-        ]
-        await asyncio.gather(*tasks)
-
-        if zconfig.NODE['id'] != zconfig.SEQUENCER['id']:
-            await asyncio.sleep(10)
-    finally:
-        zdb.pause_node.clear()
-
-
-def switch_sequencer(old_sequencer_id: str, new_sequencer_id: str):
-    """
-    Synchronous wrapper for sequencer switching.
-    """
-    with switch_lock:
-        asyncio.run(_switch_sequencer_core(old_sequencer_id, new_sequencer_id))
-
-
-async def switch_sequencer_async(old_sequencer_id: str, new_sequencer_id: str):
-    """
-    Asynchronous version of sequencer switching.
-    """
-    with switch_lock:
-        await _switch_sequencer_core(old_sequencer_id, new_sequencer_id)
-
-
-async def sync_app_finalized_state(app_name: str, new_sequencer_id: str):
-    """
-    Process a single app during sequencer switch by updating its finalized batch records
-    and resetting batch timestamps.
-
-    Args:
-        app_name: Name of the app to process
-        new_sequencer_id: ID of the new sequencer
-    """
-    all_nodes_last_finalized_batch_record = await find_all_nodes_last_finalized_batch_record_async(app_name)
-    if all_nodes_last_finalized_batch_record:
-        zdb.reinitialize(app_name, new_sequencer_id, all_nodes_last_finalized_batch_record)
-    zdb.reset_not_finalized_batches_timestamps(app_name)
